Update layout, text and remove unneeded files
This commit makes substantial changes to the text and layout of the home and contact pages. Alongside it, unnecessary files including AboutPage, InfoCard, InfoText, and the SVG logo were deleted. The text of the README.md file was also updated. Lastly, padding to the left and right was added in the index.css file to adjust layout.