Add logo to homepage and update project setup
This update introduces logo image to the HomePage and ContactPage. The logo is'centered and takes 50% of the column's width. In addition to this, setup files are modified to include specific project settings and service account creation in common projects. The project ID in setup.md is updated, and a new command to set the project for gcloud is added.